Add a config setting for the console log level.
Previously, if console logging was enabled, then the log threshold
level was hard-coded to 'info'.

This change both makes that configurable, and changes the default
from 'info' (which is very noisy) to 'warn' (which should only
include things the user really needs to know).

This is part of the work to address #1157