Configure StackDriver logging for Datalab instances.
Previously, Datalab instances wrote a very small amount of
information to StackDriver logging, and then went silent.

With this change, the logs from the Datalab server that are
higher than a configurable threshold will be written to StackDriver
logging.

This change also fixes some badly-named SSH flags that are
used to configure the verbosity of the SSH output to indicate
that those flags are specific to just the SSH connection.

This fixes #1157