Do not try to launch the browser if it may not support Datalab.
This change stops the CLI from automatically launching the browser
if the type of browser launched is one that may not work with the
Datalab UI.

In particular, this detects a large class of text-only browsers,
including the one that is provided in Google Cloud Shell.

This mitigates iessue #1139, but does not completely fix it, as
we still do not tell the user to launch the web-preview feature.

However, even though this approach still leaves room for improving
the usage under Cloud Shell, it also greatly improves the experience
when running under other environments where a text-only browser
would be opened instead of a browser with a GUI.