Prevent users from accidentally sharing a Datalab instance.
The overall goal of this change is to make it clear to users that
the Datalab instances they create should be tied to a single user.

We do this by recording for which user the Datalab instance was
created, and then checking that when a user tries to connect to
the instance.

By default, the Datalab instance is created for the user who runs
the `datalab create` command, but there is a new `--for-user` flag
that can be used to override that.

When a user tries to connect to a Datalab instance that was created
for someone else, the code prints out an error message explaining
the user mismatch.

The user still has the option of disabling that check by passing
in the new `--no-user-checking` flag, but this should prevent
users from accidentally stomping on each others' toes.

This fixes #1165