Detect when the CLI is run from Google Cloud Shell
This change teaches the CLI to detect when it is running inside
of the Google Cloud Shell, and then handle successful connections
differently for that case.

The purpose of this is that the user cannot connect to the 'local'
port in Cloud Shell via the 'localhost' domain name, but rather
must use the 'Web Preview' feature of Cloud Shell to connect to
their Datalab instance. As such, the tool should tell the user
to connect using the 'Web Preview' feature.

This fixes #1139