Replace 127.0.0.1's with localhost to support IPv6-only machines.
When running on machines with no IPv4 interface, 127.0.0.1 is not a useful IP or
interface name.  localhost on the other hand is ~universally useful for the same
concept, and works equally well on each of IPv4-only, IPv6-only, and dual-stack
machines.

This change replaces 3 explicit usages of "127.0.0.1" and two implicit ones
(hiding behind the default or null arguments to tcp-port-used's functions.

b/35224793