Make the volume mounts for Datalab instances more sensible
This change alters the way that the persistent disk used for
storing notebooks is exposed to the Datalab container.

Rather than the entire disk being mounted by the container at
`/content`, we just mount the `datalab` subdirectory at
`/content/datalab`.

That change allows us to create a `tmp` subdirectory, and mount
that into the container at `/tmp`. That, in turn, prevents temp
files created in the container from filling up the VM's boot disk.

This fixes #1192