Don't combine stderr and stdout when calling gcloud
This change fixes calls to gcloud where the same file was being
used for both the stdout and stderr of the subprocess call.

That caused issues where we needed to parse the stdout, but text
written to stderr was being intermixed with it. It also caused
more subtle issues where gcloud would write something to stderr,
and we would report it to the user via stdout (instead of stderr).

There is only one spot left where both stdout and stderr are written
to the same file, and that is where we are simply logging both of
them.

This fixes #1197