Fix the program name in help output.
This change fixes the output of the assorted `datalab --help` and
`datalab <command> --help` invocations, so that the usage string is
printed with the command name always being `datalab`.

The reason for this is that the way the tool is bundled with the
Cloud SDK results in the arg[0] value being the fully qualified
path to the 'datalab.py' file, rather than being the command that
the user entered.

That, in turn, results in the printed usage string being overly
verbose.