Quote command names in help messages.
This change makes the help messages more readable by quoting the
reference to the command that is being described.

For example, the help output for the create command previously said

    datalab create creates ...

but the new output says

    `datalab create` creates ...

The result is that it is easier for the reader to distinguish
between the reference to the command invocation and the description
of what that command does.