Remove the ensure_sshable method.
Previously, the `datalab connect` command was calling `gcloud compute ssh`
twice, once with output forwarded to the user, and once with output
written to a log file.

That was causing issues when a user has a passphrase on their SSH key,
because they would be prompted for the passphrase repeatedly.

This also didn't seem to save anything in terms of the actual verbosity
of the `connect` command output.

This fixes #1214 (at least as much as we can).