Fix intermittent undefined notebook error (#1243)
It's not clear to me why this happens from time to time, it might be that the reference to notebook maintained by the closure gets changed somehow, and some uses don't even have a closure to keep the reference. Using the global object Jupyter.notebook is safe in this context, and fixes this issue.

This also fixes the keyboard shortcut for undoing cell deletion, which started getting assigned with this fix, but was calling a deprecated method.