Add syntax coloring for `%bq query` (#1242)
Installs google.datalab.notebook as a separate Jupyter extension in order to allow it to define its own CodeMirror modes. It's also good to make the separation now so it's easier to deprecate the old namespace datalab.

The also fixes CSS colors to make things more consistent between light and dark themes.