Don't prompt for a zone when the user specifies a non-existant instance.
Prior to this change, running a datalab command against a non-existant
instance (e.g. `datalab connect does-not-exist`) would prompt the user
to specify a zone for that instance prior to telling them that it does
not exist.

However, at that point the code already ran a list command to find
matching instances, so it should no that there are none.

This change makes the pompt a little more intelligent so that it
can bail out early if it knows there is no valid zone a user could
provide for which the instance could be found.

This change also makes changes the entrypoint so that we only print
stack traces for user errors when the `--verbosity` flag has been
set to either `debug` or `info`.

This fixes #1224