Add a settings pane (#1271)
As we're adding more features to the editor, and more user preferences, we need a standalone settings pane. This PR adds a simple one that shows as a modal using a new settings button.

This is becoming too many buttons on the top right toolbar, we might want to discuss rearrangement there.