Add "Jump to file" functionality (#1282)
This PR adds a "Jump to file" input box in the file list UI that shows a list of files matching the entered substring to help with quick navigation.

Files on the user's content directory are indexed at the start of the webserver and cached in a global file index with a cap of a million files, which are then filtered as the user types a substring. Each 500ms, a GET xhr request is sent to the server containing the characters typed by the user to filter files. Only 20 search results are sent back to the client.

Chokidar (MIT) is used to index and watch the file system for changes.

For now, exact match of substrings is used, next we can look into making search fuzzy.

I also modified tests to start 10 seconds after starting the datalab server, this helps fix an issue when indexing slows down the server for the first 30 seconds or so (although it still responds to GET requests), which was causing tests to fail.

------------------------

* started on filesearch

* remove fuzzy search from server side

* style search box

* launch selected notebook

* move find input into breadcrumb header

* remove obsolete reference to settings.css

* move to full disk index and search on server side

* catch file read errors

* move to async implementation for indexing, other fixes

* disable file search when there's too many files

* move the top '20 message' to bottom of find results

* use chokidar to index and watch fs changes

* add chokidar to the software listing

* make tests wait for 10 seconds before starting notebooks

* get find stats from server on each GET request

* add setting to disable file finder

* encode pattern parameter to escape tokens, add comments about raw and error events

* typo