Remove support for kernel gateway backends. (#1288)
* Remove support for kernel gateway backends.

This change removes the code that previously allowed Datalab to
run with a kernel gateway as its backend.

We were not able to provide a sufficiently reliable user experience
in that mode, so it was deprecated with our GA release.

Since it is deprecated, this change removes the code for supporting
it, and replaces it with an error message telling the user that it
is no longer supported. That error message also lists the last release
prior to GA in case the user does not wish to upgrade.

Since the kernel gateway image can no longer be used as the backend
for a Datalab instance, we would normally delete it. However, we
have had user requests to support running notebooks as HTTP servers
via kernel gateways, so there is another use case for them.

To that end, this change also adds the changes needed to fix #1146

* Fix gateway notebook-http mode

* Remove remaining references to KG_URL