Fix front end script load order (#1335)
This is fixing the previously arbitrary script load order, by waiting on events triggered by Jupyter for each of the four template pages.

The scripts we execute in any given page to customize it can have some code that must execute before the page is initialized by Jupyter (such as adding event listeners), but it also has some code that assumes elements have already been loaded (such as customizing functionality of buttons). This PR splits Datalab's customization scripts into preLoad and postLoad scripts, and calls them appropriately.

The big change in nb.html is just moving code around, no changes to that code were made.

The change also adds a window.datalab.loaded flag signaling that loading was finished. Later on I'll use this so that tests can wait on pages to load.

Also, it seems like the Travis environment has a deep $HOME directory that gets indexed when testing starts, so I've changed the run script for tests to create an empty directory and map it as /content instead of $HOME.