Add a terminal functionality (#1341)
This adds the Jupyter terminal functionality in Datalab. We will want to customize this for the Datalab experience, but we can do that in a later PR when we decide on a good usability design for it.

A few things I could not get working:

- Tab completion. Pressing tab just sends the cursor to the next tab stop. This is probably related to the next point.
- Terminal does not understand arrow keys. Pressing Up for example prints ^[[A. This looks like it's a limitation imposed by docker. Running stock Jupyter inside docker has the same problems.
- Coloring the prompt isn't working. Again, likely related to the point above.

I'll keep looking to see what's different about the docker environment that's causing this. For now, this is good enough for basic usage.