Fix infinite recursion in Tree page init - add test (#1348)
This PR fixes an infinite recursion in the Tree page's post initialization script, which results from calling that script whenever the notebook list is redrawn, which happens periodically. One step in that init code is to override the selection_changed function with a wrapper. If this is done more than once, we get an infinite recursion since we lose our reference to the original function.

The best fix I could find is to raise a flag after calling the init code the first time. Jupyter does not offer a better event to listen to for page completion.

This PR also adds a regression UI test that makes sure selection still works after reloading the notebook list.
Another small fix included here is for a bug in the Travis file that wouldn't print broken images after testing.

* adding test for showing/hiding extra tree buttons

* fixed test. fixed tree post-init code being called more than once

* pr comments

* use window.datalab.loaded to avoid multiple initializations