Do not set SSH options on Windows.
This change updates the `datalab` command line tool so that the
`-o` option to the `gcloud compute ssh` command is only passed
when running on Posix systems (e.g. Linux and Mac OS X).

This is necessary because some SSH clients (notably, PuTTY) do
not support that option, so passing it in can cause the SSH
client to error out.

This was not an issue prior to the 154.0.0 release of the Cloud SDK,
as earlier versions would filter out the `-o` flag when running on
Windows. However, with 154.0.0 that filtering has been removed and
the `-o` option is passed in as-is.

This fixes #1356