Terminal fixes: use bash, UI improvements (#1355)
This fix adds a bashrc file inside the container to set up the enviornment for the user. Currently it changes directory to /content, colors the prompt, and adds a message to tell the user to keep sensitive work in /content. The ENV env var is set so that file gets loaded by all shells.

The rest of the PR is UI fixes to make things look better, and fixed the order of loading the Jupyter dialog module.