Fix bug in the list command that filtered new instances.
The `datalab` command line tool adds a `datalab` tag to every
instance that it creates. As such, the `datalab list` command
should only show the instances that have that tag.

Previously, `gcloud compute instances list` did not support
filtering on nested fields (such as the 'items' field of the
'tags' object), so we could not directly filter on that tag.

However, the beta labels feature (at that time) created a top
level label for every tag on an instance. This meant that we
could filter out non-datalab instances by filtering on that label.

The label feature has since been changed so that tags no longer
get copied into labels. That change results in the `datalab list`
command being too aggressive in its filtering, and the output
always being empty.

Coincidentally, the limitation on filtering on nested fields
has since been removed, so we can fix the bug by switching over
to filtering on tags (rather than labels).

This fixes #1366