Make the command line tool work well with redirected stdout.
The `datalab` command line tool prints various messages to stdout,
but it was previously never calling `sys.stdout.flush()`.

That is fine when the program is run with stdout being the user's
console, because in that case the writes would not be buffered.

However, if a user runs the tool with stdout directed to a regular
file, then the writes to stdout are buffered until `sys.stdout.flush()`
is called. That, in turn, meant that the output of `print` calls
would not be written until the program exited (which results in a flush).

This becomes an issue when a user wants to issue a long-running
invocation of the tool (such as the `connect` command) in a background
process and watch the status using something like `tail -f`.

This change fixes that issue by adding a new `print_and_flush` helper
method to the `utils` package and then using that instead of calling
`print` directly.