Move to version file backwards-compatible release.sh (#1379)
This change adds a template file for the config_local.js that gets uploaded to GCS on release. This template has placeholders for fields that are changed by the release process.

The release script is also changed to be backwards-compatible with this new format and the old one.

In addition, I added a version.sh file that contains the source of truth for the Datalab version that is used by the Datalab/Gateway build scripts as well as release script. I also bumped the version to the next semver value 1.2.DATE, ideally this should be a standalone change before each release.

* move to version file and local copy of config_local

* validate current version

* fix check for version array length