Initial pass at using Container Builder.
This commit is a first-pass at using
[Container Builder](https://cloud.google.com/container-builder/)
to build our Docker images.

For now, this is not very useful, as it only builds our base images,
which are only ever used as prerequisites for later image builds.

However, this does let us make progress towards a more complete build
by ensuring that our base images build successfully in Container Builder.