Finish flushing out the Container Builder config.
This change adds the bundling of the CLI tarball to our Container Build.

This was the last step covered by `tools/release/build.sh` that was
not covered by our Container Build, so this that conversion is complete
(although we are still using the daily build and will continue to do
so until our release process is updated).

This fixes #1400