Initial pass at having a configurable base path.
This commit is a first step towards resolving #1408. Specifically,
this adds a setting for the base path that Datalab should listen on,
cascades that setting to the other servers that Datalab starts
(Jupyter and Ungit), and then makes sure that requests using that
base path get routed correctly.

This does not fully resolve that feature request, however, because
the resources output by Datalab do not yet incorporate the base path
in their various links.