Remove dead code from the NodeJS server.
This change removes the old `login.ts` file that was previously
used when running Datalab at the command-line with a kernel gateway
backend.

The rest of the code supporting that flow was removed long ago,
but I somehow missed this file and it has just sat here unused
ever since.