Ground work for experimental UI with Polymer+TS (#1416)
This change adds the infrastructure needed to build a Polymer 2.0 front end with Typescript. It adds an environment variable DATALAB_EXPERIMENTAL_UI that can be passed to the container to turn on this UI. The new frontend resources are kept under sources/web/datalab/polymer/*, and are built into build/static/experimental/* so the old static resources aren't changed.

Two new build dependencies are introduced: bower, which is currently the main package manager used by Polymer, and polymer-cli, which is a toolbox for creating, linting, building, and serving polymer apps. These can probably be factored into a single build system later, but for now, I'm using the cli to build the static files, and optionally bundling/minifying them.

This change breaks off a piece of the UI I'm building to make code review easier, so only one element is added here: the datalab shell element, which will contain other elements introduced later.

** Typescript:
As of now, there's officially no good solution for building Polymer 2.0 apps with Typescript, so the solution I have isn't perfect. I added a type declaration file (third_party/externs/ts/polymer/polymer.d.ts) that I adapted from Polymer 1.0 that seems to be covering all the main scenarios I tested, and referenced that file in the tsconfig file so those types are seen by all elements without explicitly referencing it. I also could not find a module loader that will work with this setup, since the javascript generated by the tsc transpiler using ES6 as target (which is needed by Polymer 2.0) uses es6 modules, which aren't natively supported by any browser yet, and will need a polyfill. So for now, we'll have to reference the javascript files explicitly in the HTML of the element importing them.

* init polymer ui redo

* basic toolbar and sidebar

* add selected attributes for sidebar and button

* rename toolbar-button to sidebar-button

* use app-layout elements instead of paper-toolbar

* wrap everything in datalab-app element

* move polymer files into polymer-redo dir

* hook up click events in sidebar-button and datalab-sidebar

* strip down the sidebar to only the list of buttons, so it can be embedded anywhere

* hook up iron-pages and switch pages on client

* get rid of sidebar-button, use iron-selector, and set up client routing

* add svg logo

* added settings pane

* fix css issues with scrolling

* added help dialog

* move dialogs to top-level to avoid stacking context trap

* add iron-pages for file services

* add missing bower dependency

* using tabs instead of iron-selector, styling

* started on new design

* sidebar fixes to match new design, shared styles

* add file list header and toolbar

* adding polymer build configs

* move paper-dialogs into toolbar element, style fixes

* move polymer inside web sources directory

* start serving polymer scripts from datalab

* fix build break and load both old and new interfaces

* started on content manager

* fix idletimeout issues

* use absolute heights in preparation for item-list element

* initial item-list element, with multi-selection and highlight

* add striping and hover to item-list

* add navigation to datalab-files, still need backward navigation

* added nav icons, breadcrumbs next

* working breadcrumbs

* working backward/forward nav

* added progress bar, thinner breadcrumb borders

* add folder/file icon to list, color fixes

* clicking opens notebooks, + refresh button, refresh every 10s

* generalize item-list for both files and sessions

* working sessions page, using ApiManager

* added an env var for experimental ui

* redirect base path to /files

* cleanup

* added simple polymer ts definition file

* cleanup

* adding experimental to gitignore

* first element in typescript: datalab-app, other ts cleanup

* datalab-toolbar is now in typescript

* datalab-sidebar is now in typescript

* item-list is now in typescript

* started on ApiManager ts conversion

* rename custom_elements to components

* datalab-files and datalab-sessions are now in typescript

* complete migration to typescript

* remove most elements planning for the review

* remove redundant copy from build

* prepend base path, fix path references

* pr fixes, turn off experimental UI by default in run script