Add preview images that are built on every push to master (#1429)
This change also fixed some hard-coded bucket names so that it is possible to test the Container Builder config by pushing to a test project.

Similarly, this also changed the bucket name from where we pull the production config_local.js file so that it is hard-coded. The difference is that one is independent of the project ID, whereas the others are not.