[Polymer UI] Implement input dialog as a subclass of base dialog (#1447)
As we're adding more types of dialogs, using inheritance is a cleaner and easier approach.

In order to use inheritance with Polymer, this change adds a template stamp mechanism, which takes the subclass and baseclass, and stamps the former into an element inside the latter. This way we can reuse the Javascript across the elements, and we can extend the template at the same time.

This change adds a BaseDialog type that shows a dialog with room for a title, messag, and ok/cancel buttons. The InputDialog then extends this to add an input element with some more input behavior.

* implement input dialog as a subclass of base dialog

* missing Utils.stampInBaseTemplate method from merge

* actually clone the extended template, some style fixes

* use const