Don't invoke `gcloud auth list` for every page load. (#1449)
This change adds a cache of the gcloud account that is stored on the server
side so that it no longer needs to invoke gcloud auth list on every page load.

With this change, navigating the file directory is much faster. In my
testing, each page load of a tree/... URL takes less than 1 second.