[Polymer UI] Adding a details pane to file picker (#1444)
This change adds a toggle-able pane on the right side to the file picker. If exactly one file or notebook is selected in the picker, the pane on the right shows a preview for it. For notebooks, it fetches the first two cells, and renders any markdown in them. For any files with text/* mime types, the first 30 rows are fetched and printed. This can be expanded to show other file types as well, such as images.

* added blank details pane

* added few details to details pane

* use Polymer's array mutation methods for item-list's selectedElements

* details pane fixes, shared style for <hr>

* working details pane

* details pane styling

* add details pane placeholder div

* loading text/ and application/ mime types in details pane

* cleanup

* use marked instead of showdown for license, add marked to datalab.txt

* some pr fixes

* remove title and creation/modification dates from preview

* better placeholder text

* add preview status message

* resolve conflict with directory picker