[Polymer UI] Make design more responsive with screen width (#1484)
This change adds some responsiveness to the datalab-files element, by splitting the toolbar into two sets of buttons: Add buttons, and Update buttons. Each set collapses separately into a dropdown to save space when the screen is too narrow.

To do this, the parent datalab-app element now listens for window size changes, and calls resize handlers on the currently active page. Currently datalab-files and datalab-terminal have handlers that make use of this event.
The actual collapsing is done in javascript because it moves the buttons dynamically to an alternative dropdown element. I could not get a similar effect just using CSS, let me know if you have any ideas here.

* generate alt add toolbar on small window widths

* repeat for file ops toolbar

* change shrink sidebar threshold

* close alt toolbars when not shown

* call the resize handler when page selection changes

* move resize handler to datalab-apps to listen once

* collapse details pane on small screens

* css issue with breadcrumb refresh

* change alt toolbar background color to match original toolbars

* pr comments

* pr comments