[Polymer UI] Show an error toast on connection errors (#1509)
This PR adds a paper-toast element that shows a connection error message whenever an XHR request fails with an unexpected code. The notification is dismissed whenever requests succeed.

A new custom even NotificationEvent can be fired on any child element, and is handled by the shell element datalab-app to show/hide the notification toast. Note this is currently not used anywhere, but it's part of this PR to open the discussion about the design of notifications.

* started on paper-toast for connection errors

* show/hide connection failure toast

* remove color transition as it interferes with other animations

* cleanup

* use connected/disconnected handlers on apimanager to signal connection loss

* add signature for connection handlers

* add notification custom event, listen to it