[Polymer UI] Fix broken /files/path for download (#1516)
With this change, /files and /files/path/to/file are entirely different APIs. The first is just a view on the datalab shell entry point, the second is a file content service that's needed by Jupyter to download files. We might want to replace this for other backends than Jupyter.

* fix broken /files/path for download

* comment