[Polymer UI] Refresh on page focus (#1522)
This PR changes the way files and sessions pages are refreshed. It reduces the frequency of the auto-refresh interval to a one each minute, and adds a focus handler to trigger the refresh. This makes user operations and page switches the main trigger for refresh, rather than an auto refresh in the background.

Note we still have an issue where multiple tabs with the files page open will separately make refresh calls, so if they're on different paths, the user's last visited path will be one of these at random. The only way to fix this is to add a blur event listener on the window, which is currently not working for me.

* disable files/sessions refresh when out of focus

* remove out of focus check

* cleanup

* merge master

* make list refresh not then-able