[Polymer UI] <item-list> element tests (#1527)
This PR adds tests for the item-list element. It also adds a package.json file at the root of the experimental UI directory for npm dependencies. Later on, we can download TS types automatically. For now, this manifest is used for the current build process, and to start tests. To build the front-end, you can navigate to sources/web/datalab/polymer and run npm run build. To start tests, you can do npm test from the same directory.

Testing the item-list is done in isolation, by feeding rows and columns arrays, and checking various interactions.

One gotcha here is for webcomponent tester to play nice with the Chrome devtools page, the sauce plugin needs to be disabled, otherwise it keeps refreshing the page, which is very annoying for debugging tests. See Polymer/web-component-tester#242 (comment).

* tests for item-list

* cleanup

* use npm for build

* make wtc non-persistent, and use all system's browsers

* a few more test cases

* fix test for 5 items

* typo