[Data First] Flatten out table schema (#1540)
This PR adds support for nested schema tables in Data First page by flattening out the schema to separate rows. This is similar to the experience on bigquery.cloud.google.com. This is not necessarily the best experience, but it's a good starting point.

The PR also adds a small delay before sending API requests when the user is typing in the search box as a means of buffering.

* flatten out schema

* cleanup

* actually copy flattened fields, add test for flattened nested schema

* revert search input delay improvement

* cleanup test

* merge master

* added TODO