[Polymer UI] Divider element, use in Data First page (#1547)
This adds a vertical divider element with two panes. This element can be dropped in the document with two child divs with slots "left" and "right", and is customizable to provide the initial split percentage, and the minimum pane size.

The element is not visible by default, but the mouse pointer will change to a resize cursor when the user hovers on it, so it's up to the elements using the divider to distinguish the two panes somehow to make the divider visible if needed, for example by adding a right border to the left pane, or a box-shadow to either pane. While it's being dragged, its background color changes slightly.

Future work for this element could be making the divider customizable.

* added resizable divider element

* add slots and initial position to resizable divider

* use resizable divider in datalab-data

* use absolute positioning for divider

* merge issue

* use box-sizing for slotted elements

* cleanup

* tests for divider

* pr comments

* pr comments