Testing work for <datalab-files> element (#1549)
This PR adds some work needed to start testing the element, it does the following:

Make the element's ready method async, and return a resolved promise if it had already run so it can be called multiple times.
Remove unneeded basePath property. This wasn't actually doing anything, and the base path is being tracked and added to all requests by ApiManager.
Make the _fetchFileList method thenable so the test can wait on it.
Move some common test utilities into one file.
Add some basic tests for the files element. There are a lot more tests to add here, but this is just to get started.

* working on files test

* working on files test

* remove BaseElement, add some tests

* remove unneeded basepath

* pr comments

* remove old test name