Refactor some modules into custom elements (#1564)
This PR moves the logic from some pure javascript modules into custom elements. It also adds the abstraction of an ApiManager as a low-level means of sending xhr requests to a given backend. The rest of the ApiManager's operations are going to be moved into a different class, perhaps FileManager, in an upcoming PR.

* IApiManager abstraction

* derived ApiManager classes

* factory for creating ApiManager singleton

* using ApiManagerFactory in datalab-app

* convert ApiManager classes into custom elements

* convert SettingsManager to custom element

* convert DaasApiManager to custom element

* convert TimeoutManager to custom element

* created TerminalManager and SessionManager

* working on file manager

* fix tests

* pr comments