Add file source dropdown to file browser (#1684)
This PR adds the ability to switch the file browser's source using a dropdown. In order to achieve this, it does the following:

Adds an app setting for the list of supported file sources, which gets read on startup.
Adds icons for: local disk (jupyter), drive, bigquery, and github.
Adds a dropdown, similar to the file browser's toolbar dropdowns that appear when the window is too narrow. In this it also fixes an issue where the dropdown doesn't dismiss after clicking a button, by listening on the blur event.
Adds a SharedDriveFileManager that extends DriveFileManager and changes the gapi query appropriately.
Cleans up the FileManagerFactory's methods for getting and converting file manager types.

* added filesource dropdown, closes on blur

* file sources switcher working

* fall back to jupyter

* fix test

* only switch the type if it's changed

* cleanup

* pr comments

* className -> class

* class -> typeClass