Delete the unused `Dockerfile.py2`.
I intended to use this Dockerfile to slim down an image, but realized that
it's completely unused. As evidence, this Dockerfile has had a syntax error in
it since ba54dc6, which was over three months ago. In addition, it's clearly
diverged from the main Dockerfile, eg it was never updated for the debian ->
ubuntu switch.